rework zupass login ui (avoid using zupass libs)
